NEXUS/requirements.txt
anthropic>=0.25.0
PyQt6>=6.6.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13
requests>=2.31.0
python-dotenv>=1.0.0
pyautogui>=0.9.54
psutil>=5.9.0
screen-brightness-control>=0.23.0
pywhatkit>=5.4
Pillow>=10.0.0
opencv-python>=4.8.0
ultralytics>=8.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pyperclip>=1.8.2
wikipedia>=1.4.0
numpy>=1.24.0
